Refactor open_close + schedule + dump
I forgot to keep track of what was what, so just adding all relevant changes.
Most notable is making open_close and schedule both use the same results function instead of double.
And fixing a subtle bug around how the is_cold/is_warm was calculated as well as standardizing what
an invocation is and what a client request is.

Also, a note to self: do not bother with HA on local node. It just results in more failures.
Expect close to 100% success rate and crippling performance when HA=true.